requirements.txt
pdf2image==1.16.0
pdfminer==20191125
Pillow==9.2.0
PyPDF2==2.8.1
pytesseract==0.3.9
